Backend updates (#81)
* update prisma schema with additional profile fields. fix create-user dto duplication conflict. update create-user dto with much more detailed schema and validation for all fields. update create user route with enhanced validation, error handling, bcrypt password hashing. update other backend files in preparation for overhaul.

* update auth signup and signin logic with updated create-user dto. fix auth password comparison issue.

* move logic from controllers to services in auth and users resources. update logic to be more concise and readable.

* add purge script

* update response entities in auth and user. implement type transformation for responses.

* consolidate user response entities into single file and rename entities for better readability and correlation with routes.

* fix module import paths in auth service. update providers in auth service spec.

* update profile resource: update dtos with validation and updated fields, update entities with new fields and consolidate all response entities into one file with better more comprehensive naming, update profile controllers with new entities, move controller logic into service, add data serialization to returned profile data, update profile post logic.

+++ server/src/users/users.service.ts
@@ -1,14 +1,29 @@
-import { Injectable } from '@nestjs/common';
+import {
+  ConflictException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { PrismaService } from './../prisma/prisma.service';
+import { UserEntity } from './entities/user.entity';
+import { plainToClass } from 'class-transformer';
+import * as bcrypt from 'bcrypt';
 
 @Injectable()
 export class UsersService {
   constructor(private prisma: PrismaService) {}
 
   async create(createUserDto: CreateUserDto) {
-    return await this.prisma.user.create({
+    const userExists = await this.findUserByEmail(createUserDto.email);
+
+    if (userExists) {
+      throw new ConflictException('this email already exists');
+    }
+
+    createUserDto.password = this.hashPassword(createUserDto.password);
+
+    const createdUser = await this.prisma.user.create({
       data: {
         ...createUserDto,
         profile: {
@@ -19,49 +34,121 @@ export class UsersService {
         profile: true,
       },
     });
+
+    const serializedUser = plainToClass(UserEntity, createdUser);
+
+    return serializedUser;
   }
 
   async findAll() {
-    return await this.prisma.user.findMany({ include: { profile: true } });
+    const users = await this.prisma.user.findMany({
+      include: { profile: true },
+    });
+
+    if (!users.length) {
+      throw new NotFoundException('No users found');
+    }
+
+    const serializedUsers = users.map((user) => plainToClass(UserEntity, user));
+
+    return serializedUsers;
   }
 
   async findOne(id: number) {
-    return await this.prisma.user.findUnique({
+    const user = await this.prisma.user.findUnique({
       where: { id: id },
       include: { profile: true },
     });
+
+    if (!user) {
+      throw new NotFoundException(`User with id: ${id} not found`);
+    }
+
+    const serializedUser = plainToClass(UserEntity, user);
+
+    return serializedUser;
   }
 
   async findAllStudents() {
-    return await this.prisma.user.findMany({
+    const students = await this.prisma.user.findMany({
       where: { role: 'STUDENT' },
       include: { profile: true },
     });
+
+    if (!students.length) {
+      throw new NotFoundException('No students found');
+    }
+
+    const serializedStudents = students.map((student) =>
+      plainToClass(UserEntity, student),
+    );
+
+    return serializedStudents;
   }
 
   async findAllTutors() {
-    return await this.prisma.user.findMany({
+    const tutors = await this.prisma.user.findMany({
       where: { role: 'TUTOR' },
       include: { profile: true },
     });
+
+    if (!tutors.length) {
+      throw new NotFoundException('No tutors found');
+    }
+
+    const serializedTutors = tutors.map((tutor) =>
+      plainToClass(UserEntity, tutor),
+    );
+
+    return serializedTutors;
   }
 
   async update(id: number, updateUserDto: UpdateUserDto) {
-    return await this.prisma.user.update({
+    const userExists = await this.findOne(id);
+
+    if (!userExists) {
+      throw new NotFoundException(
+        `User with id: ${id} not found. Could not update.`,
+      );
+    }
+
+    if (updateUserDto.password) {
+      updateUserDto.password = this.hashPassword(updateUserDto.password);
+    }
+
+    const updatedUser = await this.prisma.user.update({
       where: { id: id },
       data: updateUserDto,
+      include: { profile: true },
     });
+
+    const serializedUser = plainToClass(UserEntity, updatedUser);
+
+    return serializedUser;
   }
 
   async remove(id: number) {
-    // await this.prisma.profile.delete({ where: { userId: id } });
-    return await this.prisma.user.delete({
+    const userExists = await this.findOne(id);
+
+    if (!userExists) {
+      throw new NotFoundException(`User with id: ${id} not found`);
+    }
+    const deletedUser = await this.prisma.user.delete({
       where: { id: id },
       include: { profile: true },
     });
+
+    const serializedUser = plainToClass(UserEntity, deletedUser);
+
+    return serializedUser;
   }
 
   async findUserByEmail(email: string) {
     return await this.prisma.user.findFirst({ where: { email: email } });
+  }
+
+  hashPassword(password: string) {
+    const saltRounds = 10;
+    return bcrypt.hashSync(password, saltRounds);
   }
 }
